+++ plugins/Jobs.md
@@ -26,9 +26,10 @@ Bedrock::Jobs is a plugin to the [Bedrock data foundation](../README.md) that ma
  * **DeleteJob( jobID )** - Removes all trace of a job.
    * *jobID* - Identifier of the job to delete
 
- * **RequeueJob( jobID )** - Removes all trace of a job.
+ * **Retry( jobID )** - Removes all trace of a job.
    * *jobID* - Identifier of the job to retry
-   * *nextRun* - The time/date on which the job should be set to run again, in "YYYY-MM-DD [HH:MM:SS]" format
+   * *delay* - (optional) Number of seconds to wait before retrying
+   * *nextRun* - (optional) The time/date on which the job should be set to run again, in "YYYY-MM-DD [HH:MM:SS]" format
    * *name* - (optional) Any arbitrary string name for this job.
    * *data* - (optional) Data to associate with this job
